UPDATE: moved skills and social props to page
