+++ src/routes/[[locale=supportedLocale]]/+page.svelte
@@ -6,10 +6,71 @@
 
 	const tBase = "page.home.";
 
+	const skills = [
+		{
+			title: $t("component.profile.skills.section.programmingLanguages"),
+			skills: ["Java", "C#", "Python", "JavaScript", "TypeScript", "php", "html", "css", "scss", "sass"],
+		},
+		{
+			title: $t("component.profile.skills.section.frameworks"),
+			skills: ["Spring boot", "ASP.NET Core", "Laravel", "Svelte", "SvelteKit", "React.js", "Next.js", "Vue.js", "Nuxt.js"],
+		},
+		{
+			title: $t("component.profile.skills.section.tools"),
+			skills: ["Git", "GitHub", "GitHub Actions", "Docker", "Figma"],
+		},
+		{
+			title: $t("component.profile.skills.section.concepts"),
+			skills: ["SOLID", "Xtreme Programming", "CI/CD", "OOP", "TDD", "DDD", "REST", "SCRUM", "Agile"],
+		},
+	];
+
+	const socials = [
+		{
+			username: "wiemanboy",
+			url: "https://github.com/wiemanboy",
+			icon: "github",
+		},
+		{
+			username: "Jarno Wieman",
+			url: "https://www.linkedin.com/in/jarnowieman/",
+			icon: "linkedin",
+		},
+		{
+			username: "wiemanboy",
+			url: "https://stackoverflow.com/users/20792492/wiemanboy",
+			icon: "stackoverflow",
+		},
+		{
+			username: "wiemanboy",
+			url: "https://www.instagram.com/wiemanboy/",
+			icon: "instagram",
+		},
+		{
+			username: "@wiemanboy",
+			url: "https://x.com/wiemanboy",
+			icon: "twitter",
+		},
+		{
+			username: "wiemanboy",
+			url: "https://steamcommunity.com/id/wiemanboy/",
+			icon: "steam",
+		},
+		{
+			username: "u/wiemanboy",
+			url: "https://www.reddit.com/user/wiemanboy/",
+			icon: "reddit",
+		},
+		{
+			username: "r/jarno",
+			url: "https://www.reddit.com/r/Jarno/",
+			icon: "reddit",
+		},
+	];
 </script>
 
 <main class="flex flex-grow flex-col h-screen">
-	<Profile class="m-3" name="Jarno Wieman" />
+	<Profile class="m-3" name="Jarno Wieman" {skills} {socials} />
 </main>
 
 <svelte:head>
